+++ cmd/serve/eyeshade.go
@@ -11,7 +11,7 @@ import (
 	_ "net/http/pprof"
 	// re-using viper bind-env for wallet env variables
 	_ "github.com/brave-intl/bat-go/cmd/wallets"
-	"github.com/brave-intl/bat-go/eyeshade"
+	eyeshade "github.com/brave-intl/bat-go/eyeshade/service"
 
 	"github.com/brave-intl/bat-go/cmd"
 	"github.com/brave-intl/bat-go/middleware"
