+++ eyeshade/service/kafka_handlers.go
