+++ eyeshade/service/router_static.go
