+++ eyeshade/service/service_kafka.go
